backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.11 
Pillow>=10.0.0
django-cors-headers==4.2.0
django-filter==23.3
drf-yasg==1.21.5
python-dateutil==2.8.2
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.2.3  
django-extensions==3.2.3
gunicorn==21.2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0